fix rkmpp code review issues
- rkmppdec: make rkmpp_export_frame always consume the MppFrame so
  error paths no longer double-free it or leak the DRM descriptor and
  tmp_frame
- rkmppdec: fix assignment binding on info_change so it only stores
  MPP's real info-change flag
- rkmppdec: honor AV_EF_EXPLODE for hardware-flagged errinfo frames
- rkmppenc: fully unwind the MPPEncFrame entry when submit fails so a
  bufferless frame is never queued for encoding or leaked
- rkmppenc: reclaim the consumed input frame's buffer on
  rkmpp_get_packet error exits; check av_frame_clone result
- rkrga: don't free frames already consumed by ff_filter_frame; check
  av_frame_clone result in submit_frame
- hwcontext_rkmpp: compute pool buffer size in size_t to avoid signed
  overflow on large frames
- pixdesc: rename NV20_PACKED descriptor to nv20_packed to stop
  hijacking av_get_pix_fmt("nv20") from AV_PIX_FMT_NV20LE/BE
- lavc/Makefile: add SKIPHEADERS for rkmpp headers to unbreak
  checkheaders without MPP installed

+++ libavutil/hwcontext_rkmpp.c
@@ -216,7 +216,7 @@ static AVBufferRef *rkmpp_drm_pool_alloc(void *opaque, size_t size)
     const int aligned_h = FFALIGN(hwfc->height * 5 / 4, 64);
 
     MppBuffer mpp_buf = NULL;
-    size_t mpp_buf_size = aligned_w * aligned_h * bits_pp / 8;
+    size_t mpp_buf_size = (size_t)aligned_w * aligned_h * bits_pp / 8;
 
     if (hwfc->initial_pool_size > 0 &&
         avfc->nb_frames >= hwfc->initial_pool_size)
